+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/utils/DBUtils.kt
@@ -1,10 +1,7 @@
 package com.fluttercandies.photo_manager.core.utils
 
-import android.content.ContentUris
 import android.content.ContentValues
 import android.content.Context
-import android.graphics.BitmapFactory
-import android.os.Environment
 import android.provider.BaseColumns._ID
 import android.provider.MediaStore
 import android.util.Log
@@ -14,7 +11,6 @@ import com.fluttercandies.photo_manager.core.entity.AssetEntity
 import com.fluttercandies.photo_manager.core.entity.FilterOption
 import com.fluttercandies.photo_manager.core.entity.AssetPathEntity
 import java.io.*
-import java.net.URLConnection
 import java.util.concurrent.locks.ReentrantLock
 import kotlin.concurrent.withLock
 
@@ -262,172 +258,6 @@ object DBUtils : IDBUtils {
         return assetEntity.path
     }
 
-    override fun saveImage(
-        context: Context,
-        image: ByteArray,
-        title: String,
-        desc: String,
-        relativePath: String?
-    ): AssetEntity? {
-        val cr = context.contentResolver
-        var inputStream = ByteArrayInputStream(image)
-        fun refreshInputStream() {
-            inputStream = ByteArrayInputStream(image)
-        }
-
-        val latLong = kotlin.run {
-            val exifInterface = try {
-                ExifInterface(inputStream)
-            } catch (e: Exception) {
-                return@run doubleArrayOf(0.0, 0.0)
-            }
-            exifInterface.latLong ?: doubleArrayOf(0.0, 0.0)
-        }
-        refreshInputStream()
-
-        val bmp = BitmapFactory.decodeStream(inputStream)
-        val width = bmp.width
-        val height = bmp.height
-        val timestamp = System.currentTimeMillis() / 1000
-        refreshInputStream()
-
-        val typeFromStream: String = if (title.contains(".")) {
-            // title contains file extension, form mimeType from it
-            "image/${File(title).extension}"
-        } else {
-            URLConnection.guessContentTypeFromStream(inputStream) ?: "image/*"
-        }
-
-        val values = ContentValues().apply {
-            put(
-                MediaStore.Files.FileColumns.MEDIA_TYPE,
-                MediaStore.Files.FileColumns.MEDIA_TYPE_IMAGE
-            )
-            put(MediaStore.MediaColumns.DISPLAY_NAME, title)
-            put(MediaStore.MediaColumns.MIME_TYPE, typeFromStream)
-            put(MediaStore.MediaColumns.TITLE, title)
-            put(MediaStore.Images.ImageColumns.DESCRIPTION, desc)
-            put(MediaStore.MediaColumns.DATE_ADDED, timestamp)
-            put(MediaStore.MediaColumns.DATE_MODIFIED, timestamp)
-            put(MediaStore.MediaColumns.DISPLAY_NAME, title)
-            put(MediaStore.MediaColumns.WIDTH, width)
-            put(MediaStore.MediaColumns.HEIGHT, height)
-            put(MediaStore.Images.ImageColumns.LATITUDE, latLong[0])
-            put(MediaStore.Images.ImageColumns.LONGITUDE, latLong[1])
-        }
-
-        val insertUri = cr.insert(
-            MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
-            values
-        ) ?: return null
-
-        // Write bytes.
-        val cursor = cr.query(
-            insertUri,
-            arrayOf(MediaStore.MediaColumns.DATA),
-            null,
-            null,
-            null
-        ) ?: return null
-        cursor.use {
-            if (it.moveToNext()) {
-                val targetPath = it.getString(0)
-                targetPath.checkDirs()
-                val outputStream = FileOutputStream(targetPath)
-                refreshInputStream()
-                outputStream.use { os -> inputStream.use { iStream -> iStream.copyTo(os) } }
-            }
-        }
-        val id = ContentUris.parseId(insertUri)
-        return getAssetEntity(context, id.toString())
-    }
-
-    override fun saveImage(
-        context: Context,
-        path: String,
-        title: String,
-        desc: String,
-        relativePath: String?
-    ): AssetEntity? {
-        val cr = context.contentResolver
-        val file = File(path)
-        var inputStream = FileInputStream(file)
-        fun refreshInputStream() {
-            inputStream = FileInputStream(file)
-        }
-
-        val latLong = kotlin.run {
-            val exifInterface = try {
-                ExifInterface(inputStream)
-            } catch (e: Exception) {
-                return@run doubleArrayOf(0.0, 0.0)
-            }
-            exifInterface.latLong ?: doubleArrayOf(0.0, 0.0)
-        }
-        refreshInputStream()
-
-        val (width, height) = try {
-            val bmp = BitmapFactory.decodeFile(path)
-            Pair(bmp.width, bmp.height)
-        } catch (e: Exception) {
-            Pair(0, 0)
-        }
-        val timestamp = System.currentTimeMillis() / 1000
-
-        val typeFromStream = URLConnection.guessContentTypeFromStream(inputStream)
-            ?: "image/${File(path).extension}"
-        refreshInputStream()
-
-        val uri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-        val dir = Environment.getExternalStorageDirectory()
-        val savePath = file.absolutePath.startsWith(dir.path)
-        val values = ContentValues().apply {
-            put(
-                MediaStore.Files.FileColumns.MEDIA_TYPE,
-                MediaStore.Files.FileColumns.MEDIA_TYPE_IMAGE
-            )
-            put(MediaStore.MediaColumns.DISPLAY_NAME, title)
-            put(MediaStore.MediaColumns.MIME_TYPE, typeFromStream)
-            put(MediaStore.MediaColumns.TITLE, title)
-            put(MediaStore.Images.ImageColumns.DESCRIPTION, desc)
-            put(MediaStore.MediaColumns.DATE_ADDED, timestamp)
-            put(MediaStore.MediaColumns.DATE_MODIFIED, timestamp)
-            put(MediaStore.MediaColumns.DISPLAY_NAME, title)
-            put(MediaStore.Images.ImageColumns.LATITUDE, latLong[0])
-            put(MediaStore.Images.ImageColumns.LONGITUDE, latLong[1])
-            put(MediaStore.MediaColumns.WIDTH, width)
-            put(MediaStore.MediaColumns.HEIGHT, height)
-            if (savePath) {
-                put(MediaStore.MediaColumns.DATA, path)
-            }
-        }
-
-        val contentUri = cr.insert(uri, values) ?: return null
-        val id = ContentUris.parseId(contentUri)
-        // Ignore the existence of the file, it'll be exported later.
-        val assetEntity = getAssetEntity(context, id.toString(), checkIfExists = false)
-        if (!savePath) {
-            val tmpPath = assetEntity!!.path
-            tmpPath.checkDirs()
-            val tmpFile = File(tmpPath)
-            val targetPath = "${tmpFile.parent}/$title"
-            val targetFile = File(targetPath)
-            if (targetFile.exists()) {
-                throw IOException("save target path is ")
-            }
-            tmpFile.renameTo(targetFile)
-            val updateDataValues = ContentValues().apply {
-                put(MediaStore.MediaColumns.DATA, targetPath)
-            }
-            cr.update(contentUri, updateDataValues, null, null)
-            val outputStream = FileOutputStream(targetFile)
-            outputStream.use { os -> inputStream.use { it.copyTo(os) } }
-            assetEntity.path = targetPath
-        }
-        cr.notifyChange(contentUri, null)
-        return assetEntity
-    }
-
     override fun copyToGallery(context: Context, assetId: String, galleryId: String): AssetEntity? {
         val (currentGalleryId, _) = getSomeInfo(context, assetId)
             ?: throw RuntimeException("Cannot get gallery id of $assetId")
@@ -619,71 +449,6 @@ object DBUtils : IDBUtils {
             val galleryPath = File(path).parentFile?.absolutePath ?: return null
             return GalleryInfo(galleryPath, galleryId, name)
         }
-    }
-
-    override fun saveVideo(
-        context: Context,
-        path: String,
-        title: String,
-        desc: String,
-        relativePath: String?
-    ): AssetEntity? {
-        path.checkDirs()
-        val inputStream = FileInputStream(path)
-        val cr = context.contentResolver
-        val timestamp = System.currentTimeMillis() / 1000
-        val dir = Environment.getExternalStorageDirectory()
-        val savePath = File(path).absolutePath.startsWith(dir.path)
-
-        val typeFromStream = URLConnection.guessContentTypeFromStream(inputStream)
-            ?: "video/${File(path).extension}"
-        val info = VideoUtils.getPropertiesUseMediaPlayer(path)
-        val values = ContentValues().apply {
-            put(
-                MediaStore.Files.FileColumns.MEDIA_TYPE,
-                MediaStore.Files.FileColumns.MEDIA_TYPE_VIDEO
-            )
-            put(MediaStore.Video.VideoColumns.DESCRIPTION, desc)
-            put(MediaStore.MediaColumns.TITLE, title)
-            put(MediaStore.MediaColumns.DISPLAY_NAME, title)
-            put(MediaStore.MediaColumns.MIME_TYPE, typeFromStream)
-            put(MediaStore.MediaColumns.DATE_ADDED, timestamp)
-            put(MediaStore.MediaColumns.DATE_MODIFIED, timestamp)
-            put(MediaStore.MediaColumns.DATE_TAKEN, timestamp * 1000)
-            put(MediaStore.MediaColumns.DURATION, info.duration)
-            put(MediaStore.MediaColumns.WIDTH, info.width)
-            put(MediaStore.MediaColumns.HEIGHT, info.height)
-            if (savePath) {
-                put(MediaStore.MediaColumns.DATA, path)
-            }
-        }
-
-        val uri = MediaStore.Video.Media.EXTERNAL_CONTENT_URI
-        val contentUri = cr.insert(uri, values) ?: return null
-        val id = ContentUris.parseId(contentUri)
-        val assetEntity = getAssetEntity(context, id.toString(), checkIfExists = false)
-        if (savePath) {
-            inputStream.close()
-        } else {
-            val tmpPath = assetEntity!!.path
-            tmpPath.checkDirs()
-            val tmpFile = File(tmpPath)
-            val targetPath = "${tmpFile.parent}/$title"
-            val targetFile = File(targetPath)
-            if (targetFile.exists()) {
-                throw IOException("Save target path is ")
-            }
-            tmpFile.renameTo(targetFile)
-            val updateDataValues = ContentValues().apply {
-                put(MediaStore.MediaColumns.DATA, targetPath)
-            }
-            cr.update(contentUri, updateDataValues, null, null)
-            val outputStream = FileOutputStream(targetFile)
-            outputStream.use { os -> inputStream.use { it.copyTo(os) } }
-            assetEntity.path = targetPath
-        }
-        cr.notifyChange(contentUri, null)
-        return assetEntity
     }
 
     private data class GalleryInfo(val path: String, val galleryId: String, val galleryName: String)